Authorization

+++ server/src/schema/mutations/actor.ts
@@ -7,6 +7,8 @@ import {
     AddActorToMovieInputType,
 } from "../types/actor";
 import { ActorModel } from "../../data/actor";
+import { IAuthContext } from "../../data/user";
+import { contextService } from "../../services/contextService";
 
 export const actorMutations = {
     addActorToMovie: {
@@ -16,7 +18,9 @@ export const actorMutations = {
                 type: new GraphQLNonNull(AddActorToMovieInputType),
             },
         },
-        resolve: (rootValue, { input }) => {
+        resolve: (rootValue, { input }, context: IAuthContext) => {
+            contextService.checkIsAdmin(context);
+
             const newModel = new ActorModel(input);
             const newObj = newModel.save();
             if(!newObj) {
@@ -32,7 +36,9 @@ export const actorMutations = {
                 type: new GraphQLNonNull(GraphQLID),
             },
         },
-        resolve: (rootValue, { actorId }) => {
+        resolve: (rootValue, { actorId }, context: IAuthContext) => {
+            contextService.checkIsAdmin(context);
+
             const removed = ActorModel.findByIdAndRemove(actorId).exec();
             if (!removed) {
                 throw new Error('Remove actor from movie error')
