+++ server/src/schema/mutations/person.ts
@@ -9,6 +9,8 @@ import {
 import { PersonModel } from "../../data/person";
 import { ActorModel } from "../../data/actor";
 import { CallbackError } from "mongoose";
+import {IAuthContext} from "../../data/user";
+import {contextService} from "../../services/contextService";
 
 export const personMutations = {
     createPerson: {
@@ -18,7 +20,9 @@ export const personMutations = {
                 type: new GraphQLNonNull(CreatePersonInputType),
             },
         },
-        resolve: (rootValue, { input }) => {
+        resolve: (rootValue, { input }, context: IAuthContext) => {
+            contextService.checkIsAdmin(context);
+
             const newModel = new PersonModel(input);
             const newObj = newModel.save();
             if(!newObj) {
@@ -34,10 +38,13 @@ export const personMutations = {
                 type: new GraphQLNonNull(GraphQLID),
             },
         },
-        resolve: (rootValue, { personId }) => {
+        resolve: (rootValue, { personId }, context: IAuthContext) => {
             if(!personId) {
                 throw Error("Person id is required");
             }
+
+            contextService.checkIsAdmin(context);
+
             ActorModel.deleteMany({ personId: personId }, (error: CallbackError) => {
                 console.error(error.message);
             });
